InstaLoader/middlewares/check_sub.py
import logging
from aiogram import types
from aiogram.dispatcher.handler import CancelHandler
from aiogram.dispatcher.middlewares import BaseMiddleware
from aiogram.types import InlineKeyboardMarkup, InlineKeyboardButton
from data.config import CHANNELS
from utils.misc import subcriptChanel
from loader import bot

# print(CHANNELS)


class BigBrother(BaseMiddleware):
    async def on_pre_process_update(self, update: types.Update, data: dict):
        if update.message:
            user = update.message.from_user.id
            if update.message.text in ['/start', '/help']:
                return
        elif update.callback_query:
            user = update.callback_query.from_user.id
            if update.callback_query.data == "check_subs":
                return
        else:
            return

        result = "Botdan foydalanish uchun quyidagi kanallarga obuna bo'ling:\n"
        final_status = True
        keyboard = InlineKeyboardMarkup()

        for channel in CHANNELS:
            status = await subcriptChanel.check(user_id=user, channel=channel)
            final_status *= status
            channel = await bot.get_chat(channel)
            if not status:
                invite_link = await channel.export_invite_link()
                keyboard.add(InlineKeyboardButton(text=channel.title, url=invite_link))

        if not final_status:
            await update.message.answer(result, reply_markup=keyboard, disable_web_page_preview=True)
            raise CancelHandler()
