InstaLoader/requirements.txt
aiogram==2.25
aiohttp==3.8.6
aiosignal==1.3.1
anyio==4.4.0
api-display-purposes==0.0.3
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
Babel==2.9.1
beautifulsoup4==4.12.3
build==1.0.3
CacheControl==0.13.1
certifi==2023.7.22
cffi==1.16.0
chardet==3.0.4
charset-normalizer==3.3.0
clarifai==10.5.3
clarifai-grpc==10.5.4
cleo==2.1.0
configparser==3.8.1
contextlib2==21.6.0
crashtest==0.4.1
cryptography==41.0.4
distlib==0.3.7
Django==5.0.6
django-jazzmin==3.0.0
dulwich==0.21.6
EasyProcess==1.1
emoji==2.12.1
environs==8.0.0
fastjsonschema==2.18.1
filelock==3.12.2
frozenlist==1.4.1
future==1.0.0
googleapis-common-protos==1.63.2
greenlet==3.0.3
grpcio==1.64.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.4
importlib-metadata==6.8.0
inquirerpy==0.3.4
installer==0.7.0
instaloader==4.11
jaraco.classes==3.3.0
jeepney==0.8.0
jsonschema==2.6.0
keyring==24.2.0
magic-filter==1.0.12
markdown-it-py==3.0.0
marshmallow==3.21.3
mdurl==0.1.2
MeaningCloud-python==2.0.0
more-itertools==10.1.0
msgpack==1.0.7
multidict==6.0.5
numpy==1.26.4
outcome==1.3.0.post0
packaging==23.2
pexpect==4.8.0
pfzy==0.3.4
pillow==10.4.0
pipenv==2023.11.17
pkginfo==1.9.6
platformdirs==3.11.0
playwright==1.44.0
plyer==2.1.0
poetry==1.7.1
poetry-core==1.8.1
poetry-plugin-export==1.6.0
prompt_toolkit==3.0.47
protobuf==3.20.3
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pycparser==2.21
pyee==11.1.0
Pygments==2.18.0
pyproject_hooks==1.0.0
PySocks==1.7.1
python-dotenv==1.0.1
python-rapidjson==1.18
python-telegram-bot==21.3
pytube==15.0.0
pytz==2024.1
PyVirtualDisplay==3.0
PyYAML==6.0.1
rapidfuzz==3.5.2
regex==2024.5.15
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.7.1
schema==0.7.5
SecretStorage==3.3.3
selenium==4.22.0
semantic-version==2.10.0
setuptools-rust==1.9.0
shellingham==1.5.3
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.5.0
tabulate==0.9.0
TikTokApi==6.3.0
tomlkit==0.12.1
tqdm==4.66.4
trio==0.25.1
trio-websocket==0.11.1
tritonclient==2.47.0
trove-classifiers==2023.9.19
typing_extensions==4.12.2
urllib3==2.0.6
virtualenv==20.25.0
wcwidth==0.2.13
webdriverdownloader==1.1.0.4
websocket-client==1.8.0
wsproto==1.2.0
yarl==1.9.4
youtube-dl==2021.12.17
zipp==3.17.0
